Changed structure of the map, added food and movement
Map has been changed from containing letters to numbers. Added feature of movement trough out the map and as well as eating food. Also added restriction to move trough the walls and player can now port from top of the map to the bottom and the opposite. Lighting was started, but still needs to be worked on.